// src/types.ts
export type SearchFormText = {
  headerText: string;
  descriptionText: string;
  departInputPlaceholder: string;
  returnInputPlaceholder: string;
  searchButtonText: string;
};

export type SearchFormTheme = {
  backgroundColor: string;
  color: string;
  buttonBgColor: string;
  buttonColor: string;
};